Finish Tuple Conversion
All aggregate types are now tuple types.  This involved widespread
changes including a reorganization of the print infrastructure.
The main printer class is now in the core library so debug code
may use it.  The print filter is now a simple wrapper around
that core library implementation.

This also includes a fix for make dependencies.  PIC objects
previously had dependency files with the same name as non-PIC
objects, causing a race condition and incomplete dependencies.